requirements.txt
#
# This file is autogenerated by pip-compile with python 3.10
# To update, run:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via requests
cryptography==38.0.3
    # via -r requirements.in
idna==3.4
    # via requests
pycparser==2.21
    # via cffi
pyjwt==2.6.0
    # via -r requirements.in
requests==2.28.1
    # via -r requirements.in
urllib3==1.26.12
    # via requests
